+++ changelog.md
@@ -7,6 +7,7 @@
 - Feat: Non linear pipeline
 - Feat: Replace is now a valid operation
 - Breaking change: Should use $thing instead of $entity and $relation;
+- Breaking change: Mutation result are always array (even if nothing is nested) 
 - Refacto: Preparing for multiple DBs
 - Chore: Immer downgraded to 9.0.21 (until <https://github.com/immerjs/immer/issues/1087> is fixed)